adding tests to check the correctness of the bench results
